Refactor MultiCastDelegate.h to MulticastDelegate.h
