+++ Goknar/Source/Goknar/Delegates/MulticastDelegate.h
