+++ Goknar/Source/Goknar/Delegates/stdafx.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "targetver.h"
 #include "stdafx.h"
-#include "MultiCastDelegate.h"
+#include "MulticastDelegate.h"
 #include <iostream>
 #include <chrono>
